Update shortcut routes
